// ichub-frontend/src/features/industry-core-kit/part-discovery/components/submodel-addons/battery-pass-digital-nameplate/types.ts
export interface PhoneEntity {
  TelephoneNumber: MultiLangEntry | MultiLangEntry[];
  TypeOfTelephone?: string;
  AvailableTime?: MultiLangEntry | MultiLangEntry[];
}

export interface FaxEntity {
  FaxNumber: MultiLangEntry | MultiLangEntry[];
  TypeOfFaxNumber?: string;
}

export interface EmailEntity {
  EmailAddress: string;
  PublicKey?: MultiLangEntry | MultiLangEntry[];
  TypeOfEmailAddress?: string;
  TypeOfPublicKey?: MultiLangEntry | MultiLangEntry[];
}

export interface IpCommunicationChannel {
  AddressOfAdditionalLink: string;
  TypeOfCommunication?: string;
  availableTime?: MultiLangEntry[];
}

export interface AddressInformation {
  RoleOfContactPerson?: string;
  NationalCode: MultiLangEntry[];
  Languages?: string[];
  TimeZone?: string;
  CityTown: MultiLangEntry[];
  Company?: MultiLangEntry[];
  Department?: MultiLangEntry[];
  Phone?: PhoneEntity;
  Fax?: FaxEntity;
  Email?: EmailEntity;
  IPCommunicationChannels?: IpCommunicationChannel[];
  Street: MultiLangEntry[];
  ZipCode: MultiLangEntry[];
  POBox?: MultiLangEntry[];
  ZipCodeOfPOBox?: MultiLangEntry[];
  StateCounty?: MultiLangEntry[];
  NameOfContact?: MultiLangEntry[];
  FirstName?: MultiLangEntry[];
  MiddleNames?: MultiLangEntry[];
  Title?: MultiLangEntry[];
  AcademicTitle?: MultiLangEntry[];
  FurtherDetailsOfContact?: MultiLangEntry[];
  AddressOfAdditionalLink?: string;
}

export interface MarkingFile {
  value: string;
  contentType: string;
}

export interface Marking {
  MarkingName: string;
  DesignationOfCertificateOrApproval?: string;
  IssueDate?: string;
  ExpiryDate?: string;
  MarkingFile: MarkingFile;
  MarkingAdditionalText?: string;
}

export type BatteryLifeCycleStage =
  | 'original'
  | 'repurposed'
  | 're-used'
  | 'remanufactured'
  | 'waste';

export interface BatteryPassDigitalNameplate {
  URIOfTheProduct: string;
  ManufacturerName: MultiLangEntry[];
  AddressInformation: AddressInformation;
  SerialNumber: string;
  DateOfManufacture: string;
  DateOfPuttingIntoService?: string;
  UniqueFacilityIdentifier: string;
  LifeCycleStage: BatteryLifeCycleStage;
  OperatorIdentifier?: string;
  ManufacturerIdentifier: string;
  Markings: Marking[];
  EUDeclarationOfConformity: string[];
  ResultsOfTestReportsProvingCompliance: string[];
}